+++ extras/src/lib.rs
@@ -1,0 +1,9 @@
+#![doc = include_str!("../README.md")]
+#![deny(clippy::all)]
+#![deny(clippy::pedantic)]
+#![allow(clippy::cast_possible_wrap)]
+#![warn(missing_docs)]
+#![forbid(unsafe_code)]
+#![cfg_attr(docsrs, feature(doc_cfg))]
+
+pub mod extensions;