+++ integrations/axum/src/subscription.rs
@@ -7,6 +7,7 @@ use async_graphql::{
         ALL_WEBSOCKET_PROTOCOLS, DefaultOnConnInitType, DefaultOnPingType, WebSocketProtocols,
         WsMessage, default_on_connection_init, default_on_ping,
     },
+    runtime::TokioTimer,
 };
 use axum::{
     Error,
@@ -288,7 +289,7 @@ where
                 .connection_data(self.data)
                 .on_connection_init(self.on_connection_init)
                 .on_ping(self.on_ping.clone())
-                .keepalive_timeout(self.keepalive_timeout)
+                .keepalive_timeout(TokioTimer::default(), self.keepalive_timeout)
                 .map(|msg| match msg {
                     WsMessage::Text(text) => Message::Text(text.into()),
                     WsMessage::Close(code, status) => Message::Close(Some(CloseFrame {