+++ tests/lifetime.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::diverging_sub_expression)]
+#![allow(dead_code)]
 
 use async_graphql::*;
 use static_assertions_next::_core::marker::PhantomData;